pyproject.toml
[project]
name = "isn20261"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pulumi>=3.0.0,<4.0.0",
    "pulumi-aws>=7.0.0,<8.0.0",
    "pulumi-aws-apigateway>=3.0.0,<4.0.0",
    "pulumi-awsx>=3.0.0,<4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "moto[cognitoidp,dynamodb]>=5",
    "pyjwt>=2",
    "cryptography>=44",
]
